Removed redundancy when assigning robobee skin IDs
Skin IDs in DST are generated by simply indexing every skin for a prefab and using the skins index as its ID.

The original implementation took advantage of this by simply adding the prefab and its skins to the skin list and then regenerated all skin IDs in the entire game using a loop, similar to the one in prefabskins.lua.

This new implementation directly sets the robobee skin IDs directly and leaves the default IDs untouched.